requirements.txt
openai>=1.0.0
pillow
requests
python-dotenv
cloudinary
schedule
textblob
